requirements.txt
Babel==2.16.0
langchain_core==0.3.28
langchain_openai==0.2.14
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
